update: Added some more currency tests
